Bmi changed

+++ BmiService.java
@@ -1,7 +1,7 @@
 public class BmiService {
     public double calculate(double weight, double height) {
         double bmi = weight / (height * height / 100 / 100);
-              return (int) bmi;
+              return (double) bmi;
     }
 }
